:art: Moved addCourse sections to a new ts file since home.ts started getting too long
