+++ endpoints/users.json
@@ -95,7 +95,8 @@
       "phone": "0785045454",
       "address": "anjdnaskdjasn",
       "courses": [
-        "1"
+        "1",
+        "7"
       ],
       "role": "admin",
       "authToken": "2nqy7reb-ab6j-808l-z614-9izi87ytt3uj",
