+++ public/js/authentication.js
@@ -166,7 +166,6 @@ export let signUpUser = (formElement) => __awaiter(void 0, void 0, void 0, funct
     };
     let postResponse = yield fetch("http://localhost:3001/users", options);
     let postOutput = yield postResponse.json();
-    console.log(postOutput);
     location.href = "/pages/home.html";
 });
 // Function that hashes a password using SHA-256 and a salt.
